created access token

+++ Backend/controllers/expenseController.js
@@ -4,8 +4,8 @@ const addExpense = async(req, res) => {
     try {
         const{amount, description, category} = req.body;
 
-        const data = await Expense.create({amount, description, category});
-        res.status(201).json({newExpenseDetail: data});
+        const data = await Expense.create({amount, description, category, userId: req.user.id});
+        res.status(201).json({newExpenseDetail: data, success: true});
     } catch(err) {
         res.status(500).json(err);
     }
@@ -13,11 +13,11 @@ const addExpense = async(req, res) => {
 
 const getExpenses = async(req, res) => {
     try {
-        const expenses = await Expense.findAll();
-        res.status(200).json({allExpensesDetails: expenses})
+        const expenses = Expense.findAll({ where : { userId: req.user.id}});
+        res.status(200).json({newExpenses: expenses, success: true})
     } catch(error) {
         console.log('Get expenses is failing', JSON.stringify(error))
-        res.status(500).json({error: error})
+        res.status(500).json({success: false, error: error})
     }
 }
 
@@ -28,7 +28,7 @@ const deleteExpense = async (req, res) => {
             return res.status(400).json({err: 'ID is missing'})
         }
         const expenseId = req.params.id;
-        await Expense.destroy({where: {id: expenseId}});
+        await Expense.destroy({where: {id: expenseId, userId: req.user.id}});
         res.sendStatus(200);
     } catch(err) {
         console.log(err)
